Designed leaderboard

+++ app/src/main/java/com/example/login/model/User.kt
@@ -1,3 +1,3 @@
 package com.example.login.model
 
-data class User(var userId:String = "", var userName:String = "", var profileImage:String = "", var userPoints:Int = 0)
+data class User(var userId:String = "", var userName:String = "", var profileImage:String = "", var userPoints:Long = 0)
